refactor(iap): inline environment query options
