+++ apps/desktop/src/mobile/use-app-store-environment.ts
@@ -1,9 +1,14 @@
 import { useCallback } from 'react'
-import { useQuery, useQueryClient } from '@tanstack/react-query'
+import { queryOptions, useQuery, useQueryClient } from '@tanstack/react-query'
+import { getAppStoreEnvironment } from '@reflect/core'
 import { useBridgeReady } from '@/hooks/use-bridge-ready'
 import { queryKeys } from '@/lib/query-client'
-import type { AppStoreEnvironment } from '@/mobile/iap-storage'
-import { createAppStoreEnvironmentQueryOptions } from '@/mobile/iap-query-options'
+import {
+  appStoreEnvironmentSchema,
+  readIapEnvironmentSeed,
+  writeIapEnvironmentSeed,
+  type AppStoreEnvironment,
+} from '@/mobile/iap-storage'
 import { useGraph } from '@/providers/graph-provider'
 
 export type { AppStoreEnvironment } from '@/mobile/iap-storage'
@@ -18,10 +23,23 @@ export function useAppStoreEnvironment(): {
   const { platform } = useGraph()
   const bridgeReady = useBridgeReady()
   const queryClient = useQueryClient()
-  const query = useQuery({
-    ...createAppStoreEnvironmentQueryOptions(),
-    enabled: platform === 'ios' && bridgeReady,
-  })
+  const query = useQuery(
+    queryOptions({
+      queryKey: queryKeys.iap.environment,
+      queryFn: async () => {
+        const parsed = appStoreEnvironmentSchema.safeParse(await getAppStoreEnvironment())
+        const environment = parsed.success ? parsed.data : null
+        writeIapEnvironmentSeed(environment)
+        return environment
+      },
+      initialData: readIapEnvironmentSeed,
+      initialDataUpdatedAt: 0,
+      staleTime: Infinity,
+      refetchOnMount: (environmentQuery) =>
+        environmentQuery.state.dataUpdatedAt === 0 ? 'always' : false,
+      enabled: platform === 'ios' && bridgeReady,
+    }),
+  )
   const invalidate = useCallback(() => {
     void queryClient.invalidateQueries({ queryKey: queryKeys.iap.environment })
   }, [queryClient])
